move tag tests to a file
